refactor: minor sequencer cleanup
