+++ yarn-project/sequencer-client/src/block_builder/circuit_block_builder.test.ts
@@ -19,8 +19,8 @@ import { MockProxy, mock } from 'jest-mock-extended';
 import { default as levelup } from 'levelup';
 import flatMap from 'lodash.flatmap';
 import { default as memdown } from 'memdown';
-import { makeEmptyTx, makeEmptyUnverifiedData } from '../deps/tx.js';
-import { VerificationKeys, getVerificationKeys } from '../deps/verification_keys.js';
+import { makeEmptyTx, makeEmptyUnverifiedData } from '../mocks/tx.js';
+import { VerificationKeys, getVerificationKeys } from '../mocks/verification_keys.js';
 import { EmptyProver } from '../prover/empty.js';
 import { Prover } from '../prover/index.js';
 import { Simulator } from '../simulator/index.js';
