+++ yarn-project/sequencer-client/src/index.ts
@@ -4,7 +4,7 @@ import { PublisherConfig, TxSenderConfig } from './publisher/config.js';
 export * from './sequencer/index.js';
 export * from './publisher/index.js';
 export * from './client/index.js';
-export * from './deps/tx.js';
-export * from './deps/verification_keys.js';
+export * from './mocks/tx.js';
+export * from './mocks/verification_keys.js';
 
 export type SequencerClientConfig = PublisherConfig & TxSenderConfig & SequencerConfig;
